Add Image files. Move code to Painting file, PaintingTableViewCellDelegate from PaintingModel.
