+++ ArtGallery/View Controllers/Painting.swift
@@ -7,3 +7,15 @@
 //
 
 import Foundation
+import UIKit
+
+
+struct Painting {
+    let image: UIImage
+    var isLiked: Bool
+    
+    init(image: UIImage, isLiked: Bool = false) {
+        self.image = image
+        self.isLiked = isLiked
+    }
+}
