+++ ArtGallery/View Controllers/PaintingTableViewCellDelegate.swift
@@ -7,3 +7,7 @@
 //
 
 import Foundation
+
+protocol PaintingTableViewCellDelegate: class {
+    func tappedLikeButton(on cell: PaintingTableViewCell)
+}
